Move chat to features directory
