+++ edn-react/src/features/chat/types/activityItem.ts
