+++ edn-react/src/services/ChatService.ts
@@ -1,4 +1,4 @@
 import { signal } from "@preact/signals-react";
-import { ActivityItem } from "../components/chat/types/activityItem";
+import { ActivityItem } from "../features/chat/types/activityItem";
 
 export const messages = signal([] as ActivityItem[]);
